+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Nuvola SDK Change Log
 =====================
 
+
+* check-project: Added check that `metadata.in.json` use two spaces for indentation and no trailing whitespace.
+* convert-project: Save `metadata.in.json` with correct indentation and no trailing whitespace.
+
 4.11.0 - May 8th, 2018
 ----------------------